No-op: rename, reorder, and move the wait helpers into the resource file
Pure presentation change with no behavior change, split out so the next commit
shows only what the migration actually does. GitHub renders a whole-file
reshuffle as hundreds of added/removed lines and a file rename as an unrelated
delete plus add, which buries the real diff -- this absorbs all of that here.

Three mechanical edits, none of which changes behavior:

1. Renames 5 symbols to the names the generator emits, and reorders the
   resource file's 14 declarations into its canonical layout.
2. git mv data_source_resource_*.go -> data_source_*.go, so it shows as a
   rename rather than a delete plus an add.
3. Moves the 4 PLATTC wait helpers out of utils_wait.go and into the resource
   file, where the generator puts them (plattcDeleteStatus renamed to
   privateLinkAttachmentConnectionDeleteStatus). This is what stops
   utils_wait.go's deletions and the resource file's additions from reading as
   two unrelated changes.

Verified null:

  cmd/structural-diff, resource:    5 removed / 5 added -- each rename is one of
                                    each, and the only two changed bodies differ
                                    solely by the renamed call sites
  cmd/structural-diff, data source: 7 identical, 0 changed/added/removed (exit 0)
  moved helpers:                    all 4 bodies byte-identical after the move
  utils_wait.go:                    0 added lines; exactly the 4 removed
